Move test utilities to a dedicated private package
