+++ packages/test-utilities/lib/constants.ts
@@ -1,0 +1,1 @@
+export const VALID_API_KEY = 'a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6'
