+++ packages/test-utilities/lib/matchers/index.ts
@@ -1,0 +1,2 @@
+import './to-have-processed-span'
+import './to-have-sent-span'
