+++ packages/test-utilities/lib/resource-attributes-source.ts
@@ -1,4 +1,4 @@
-import { ResourceAttributes } from '../../lib/attributes'
+import { ResourceAttributes } from '@bugsnag/js-performance-core'
 
 function resourceAttributesSource (): ResourceAttributes {
   return new ResourceAttributes('test', '3.4.5', 'bugsnag.performance.core', '1.2.3')
